move get_document_store
